# Use ReadOnlySpan<char> Improve performance after .netcoreapp2.1