--- a/test/SkyApm.Core.Tests/SkyApm.Core.Tests.csproj
+++ b/test/SkyApm.Core.Tests/SkyApm.Core.Tests.csproj
@@ -1,12 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFrameworks>netcoreapp2.0;netcoreapp2.1</TargetFrameworks>
+
+    <IsPackable>false</IsPackable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="BenchmarkDotNet" Version="0.*" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.0" />
+    <PackageReference Include="xunit" Version="2.4.0" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
 
     <ProjectReference Include="..\..\src\SkyApm.Core\SkyApm.Core.csproj" />
   </ItemGroup>