Please answer these questions before submitting pull request

- Why submit this pull request?
- [ ] Bug fix
- [ ] New feature provided
- [X] Improve performance

- Related issues

___
### Bug fix
- Bug description.

- How to fix?

___
### New feature or improvement
netcoreapp2.0:

| Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
|  Parse | 827.2 ns | 16.38 ns | 21.86 ns | 0.1268 |     - |     - |     536 B |

netcoreapp2.1:

| Method |     Mean |     Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|----------:|---------:|------:|------:|------:|----------:|
|  Parse | 44.51 ns | 0.8727 ns | 1.252 ns |     - |     - |     - |         - |